% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrexplore.r
\name{blt}
\alias{blt}
\title{Barrier Layer Thickness}
\usage{
blt(z, t, s, ...)
}
\arguments{
\item{z}{Vector of depths, or 3-column dataframe containing depths in first 
column, temperatures in the second column and salinities in the third 
column.}

\item{t}{Vector of temperatures of same length as \code{z}.}

\item{s}{Vector of salinities of same length as \code{z}.}

\item{...}{Other arguments passed to \code{mld}. Note that 
\code{return.index = FALSE} is fixed.}
}
\value{
The thickness of the barrier layer.
}
\description{
Calculate the thickness of the barrier layer.
}
\details{
The 'Barrier Layer' is the region between the pycnocline and 
  thermocline. This function identifies the thermocline and pycnocline 
  depths using \code{mld} and calculates the thickness of the barrier as the 
  difference the two depths.
}
\references{
Chu, Peter C., and Chenwu Fan. "Maximum angle method for 
  determining mixed layer depth from seaglider data." Journal of 
  oceanography 67.2 (2011): 219-230.
}
\seealso{
\code{\link{mld}}
}
