% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rreformat.r
\name{build_habitat_table}
\alias{build_habitat_table}
\title{Build Habitat Table}
\usage{
build_habitat_table(d, e, w)
}
\arguments{
\item{d}{The data frame containing bathymetry data. Expects columns 
\code{bedelev}, \code{dist}, \code{count}, and \code{sum}.}

\item{e}{The elevation (depth) at which to compute habitat.}

\item{w}{The water surface elevation for which to compute habitat.}
}
\value{
A data frame containing the columns \code{elev}, \code{wse}, 
  \code{count}, \code{count.littoral}, \code{count.limnetic}, 
  \code{count.epibenthic}, \code{count.sublimnetic}, and 
  \code{count.profundal}.
}
\description{
Compute habitat volume for a given depth and water surface elevation.
}
