% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rreformat.r
\name{correct_dist}
\alias{correct_dist}
\title{Correct Transect Distance Markers}
\usage{
correct_dist(d, distcol = "dist")
}
\arguments{
\item{d}{A dataframe of CTD transect data.}

\item{distcol}{The column name containing distance data to correct.}
}
\value{
The dataframe \code{d} with corrected values in column 
  \code{distcol}.
}
\description{
Correct transect distance markers in CTD transect data.
}
\details{
Distances are corrected and converted from km to m by comparing GPS 
  coordinates and transect distance markers to RRE bathymetry. Distance 
  corrections are shown below:
  \itemize{
    \item 0.3 => 400
    \item (1.1, 1.2) => 1100
    \item 1.5 => 4800 (typo)
    \item (2.4, 2.5) => 2600
    \item (3.1, 3.2, 3.5) => 3800
    \item 4.0 => 4300
    \item 4.6 => 4800
    \item 5.3 => 5700
    \item 6.4 => 6900
    \item (7.3, 7.4) => 7900
    \item 8.7 => 9300
    \item 9.5 => 10100
    \item (10.0, 10.1) => 10800
  }
}
\seealso{
read_rrectd read_rrectdgrid merge_rrectd
}
