% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rreadwrite.r
\docType{data}
\name{ctd}
\alias{ctd}
\title{CTD Cast Data}
\format{A data frame with 14 variables}
\usage{
data(ctd)
}
\description{
CTD cast data collected by the Largier Lab and processed using
  \code{rremat}. Structured as:
\enumerate{
  \item \code{date} cast date
  \item \code{dist} longitudinal distance from mouth (km)
  \item \code{surfelev} water surface elevation (m NAVD29)
  \item \code{elev} elevation (m NAVD29)
  \item \code{depth} cast depth (m)
  \item \code{ta} temperature (degrees C)
  \item \code{sa} salinity (PSU)
  \item \code{da}
  \item \code{oa} dissolved oxygen (mg/L)
  \item \code{sat} dissolved oxygen percent saturation (%)
  \item \code{fl} fluorescence
  \item \code{bt}
  \item \code{par}
  \item \code{ph} pH
}
}
\keyword{datasets}

