% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rreadwrite.r
\docType{data}
\name{ctdlocations}
\alias{ctdlocations}
\title{CTD Cast Location Data}
\format{A data frame with 5 variables}
\usage{
data(ctdlocations)
}
\description{
CTD Cast locations. Structured as:
\itemize{
  \item \code{id} CTD cast location ID
  \item \code{name} CTD cast location name
  \item \code{latitude} CTD cast location latitude
  \item \code{longitude} CTD cast location longitude
}
}
\details{
CTD cast location data was collected from the Russian River Estuary 
  Circulation and Water Quality Data Reports (2011, 2012, 2013, 2014) 
  submitted to the Sonoma County Water Agency.
}
\keyword{datasets}
