% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rreformat.r
\name{depth_from_elev}
\alias{depth_from_elev}
\title{Add Cast Depth}
\usage{
depth_from_elev(d, depthcol = "depth", elevcol = "elev",
  surfelevcol = "surfelev")
}
\arguments{
\item{d}{A dataframe of CTD transect data.}

\item{depthcol}{The column name to assign depth data to.}

\item{elevcol}{The column name containing elevation.}

\item{surfelevcol}{The column name containing water surface elevation.}
}
\value{
The dataframe \code{d} with an additional column \code{depthcol}.
}
\description{
Add cast depths to CTD transect data.
}
\details{
Depth is calculated as \code{depthcol = surfelevcol - elevcol}.
}
\seealso{
read_rrectd read_rrectdgrid merge_rrectd
}

