% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rreadwrite.r
\name{download_streamflow}
\alias{download_streamflow}
\title{Download USGS streamflow data}
\usage{
download_streamflow(f = tempfile(), begin_date, end_date, product = "00060",
  station = c("11467000", "11467200"), ...)
}
\arguments{
\item{f}{Destination file.}

\item{begin_date}{Character start date of date request in format
"YYYY-MM-DD".}

\item{end_date}{Character end date of date request in format
"YYYY-MM-DD".}

\item{product}{Data product, default is streamflow.}

\item{station}{Data station to download data from. Default is Guerneville.
Austin Creek station is second element of default value.}

\item{...}{Other arguments to pass to \code{download.file()}.}
}
\value{
The destination file path.
}
\description{
Download data from USGS web API.
}
\details{
The NOAA Tides & Currents web API limits data requests to 365 days
  or less.
}
