% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rreadwrite.r
\name{download_tides}
\alias{download_tides}
\title{Download NOAA tide data}
\usage{
download_tides(f = tempfile(), begin_date, end_date,
  product = "hourly_height", station = "9415020", datum = "MLLW",
  units = "metric", time_zone = "GMT", ...)
}
\arguments{
\item{f}{Destination file.}

\item{begin_date}{Character or integer start date of date request in format
YYYYMMDD.}

\item{end_date}{Character or integer end date of date request in format
YYYYMMDD. must be no more than 365 days later than \code{start_date}.}

\item{product}{Data product, default is hourly ocean water height.}

\item{station}{Data station to download data from. Default is Point Reyes 
Buoy.}

\item{datum}{Vertical datum that data product is references to. Default
is mean lower low water (MLLW).}

\item{units}{units used for data product. Default is metric units.}

\item{time_zone}{Time zone used in data product. Default is GMT.}

\item{...}{Other arguments to pass to \code{download.file()}.}
}
\value{
The destination file path.
}
\description{
Download data from NOAA Tides & Currents web API.
}
\details{
The NOAA Tides & Currents web API limits data requests to 365 days
  or less.
}
