% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrevaluate.r
\name{join_volume}
\alias{join_volume}
\title{Join volume information}
\usage{
join_volume(d, volumes, cellvol, joincols = c("elev", "dist", "wse"),
  countcols = c("count.total", "count.littoral", "count.surface",
  "count.subsurface", "count.profundal"), volcols = c("volume.total",
  "volume.littoral", "volume.surface", "volume.subsurface", "volume.profundal"))
}
\arguments{
\item{d}{Dataframe containing data to have volume information attached to.}

\item{volumes}{Dataframe containing cell counts. If \code{cellvol} is 
missing, must also containing the attribute "resolution" which is a 
dataframe or list with elements \code{xres}, \code{yres}, and \code{zres}.}

\item{cellvol}{Numeric volume of cells desribed by \code{volumes}.}

\item{joincols}{Columns to use when joining \code{volumes} to \code{d}.}

\item{countcol}{Column name containing cell count data in \code{volumes}.}

\item{volcol}{The name of the column containing volume data to be added to
\code{d}.}
}
\value{
The dataframe \code{d} with the additional column \code{volcol}.
}
\description{
Attach water volume information from estuary bathymetry to gridded data.
}

