% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrevaluate.r
\name{join_wse}
\alias{join_wse}
\title{Join water-surface elevation information}
\usage{
join_wse(d, ctd, joincols = c("date", "id"), wsein = "elev",
  wseout = "wse")
}
\arguments{
\item{d}{Dataframe containing data to have water surface elevation attached 
to.}

\item{ctd}{Dataframe containing ctd transects matching grids in \code{d}.}

\item{joincols}{Columns to use when joining \code{ctd} to \code{d}.}

\item{wsein}{The name of the column in \code{ctd} containing water-surface 
elevation data.}

\item{wseout}{The name of the column containing water-surface elevation data
to be added to \code{d}.}
}
\value{
The dataframe \code{d} with the additional column \code{wseout}.
}
\description{
Attach maximum water-surface elevation recorded in CTD transects to gridded 
  data.
}

