% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rrevaluate.r
\name{make_hsd}
\alias{make_hsd}
\title{Discrete Habitat Suitability Function Generator}
\usage{
make_hsd(expr, score)
}
\arguments{
\item{expr}{List of expressions evaluating the value of a parameter \code{x}.
All expressions should be mutually exclusive. See examples for how to
construct the expressions.}

\item{score}{Vector of scores assigned when a given expression evaluates as
\code{TRUE}.}
}
\value{
A function representing the habitat suitability index.
}
\description{
Generate a discrete habitat suitability function given parameter windows and
scores.
}
\examples{
windows = list(
  optimal = expression(x >= 14 & x <= 18),
  suitable = expression(x < 14 | (x >= 18 & x < 21)),
  stressful = expression(x >= 21 & x <= 25),
  unsuitable = expression(x > 25)
)
scores = c(3, 2, 1, 0)

make_hsd(windows, scores)
make_hsd(windows, names(windows))
}

