% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rrevaluate.r
\name{make_hsi}
\alias{make_hsi}
\title{Habitat Suitability Index Generator}
\usage{
make_hsi(hsc = list(), ag.fun = mean)
}
\arguments{
\item{hsc}{A list of habitat suitability curve functions.}

\item{ag.fun}{A function used for aggregating the outputs of the habitat
suitability curves.}
}
\value{
A function that calculates the Habitat Suitability Index given a
  vector or dataframe of habitat suitability parameters corresponding (in
  order) to the habitat suitability curves contained in \code{hsc}.
}
\description{
Generate a habitat suitability index function based on a set of habitat
suitability curves and an aggregation rule.
}
\examples{
f1 = make_hsc(c(5, 10, 15, 20, 25), c(0, 0.7, 1, 0.75, 0))
f2 = make_hsc(c(0, 5, 10), c(1, 0.5, 0), scoreleft = 1)
make_hsi(list(f1, f2), ag.fun = mean)
make_hsi(list(f1, f2), ag.fun = min)
}

