% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rreformat.r
\name{merge_rrectd}
\alias{merge_rrectd}
\title{Combine CTD Transects
Combine CTD transect data files into a dataframe.}
\usage{
merge_rrectd(ctd, dates = NULL, ids = NULL, timecol = "date",
  idcol = "id")
}
\arguments{
\item{ctd}{A list of dataframes of CTD transect data.}

\item{dates}{Vector of dates of each transect listed in \code{ctd}. If
\code{NULL}, will search for numeric YYYYMMDD in \code{names(ctd)}.}

\item{timecol}{The column name to be created containing the transect dates.}
}
\description{
Combine CTD Transects
Combine CTD transect data files into a dataframe.
}
\seealso{
read_rrectd read_rrectdgrid
}

