% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rreformat.r
\name{merge_rrewll}
\alias{merge_rrewll}
\title{Combine Water Level Records}
\usage{
merge_rrewll(wll, timecol = "mtime", depthcol = "depth", sitecol = "site")
}
\arguments{
\item{wll}{A list of dataframes of water level data}

\item{timecol}{The column name containing matlab numeric datetime. Matlab
times are assumed to be in "UTC" timezone.}

\item{depthcol}{The column name containing depths.}

\item{sitecol}{The column name to contain site data.}
}
\description{
Combine water level data files into a dataframe.
}

