% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rreformat.r
\name{merge_streamflow}
\alias{merge_streamflow}
\title{Combine Streamflow Records}
\usage{
merge_streamflow(sf, datetimecol = "datetime", flowcol = "discharge")
}
\arguments{
\item{sf}{A list of dataframes of streamflow data.}

\item{datetimecol}{Column name containing timestamps.}

\item{flowcol}{Column name containing discharge.}
}
\value{
A single dataframe of streamflow data.
}
\description{
Combine streamflow records obtained from the USGS.
}

