% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rreformat.r
\name{merge_tides}
\alias{merge_tides}
\title{Combine Tide Records}
\usage{
merge_tides(tides, datetimecol = "Date.Time", heightcol = "Water.Level",
  sigmacol = "Sigma")
}
\arguments{
\item{tides}{A list of dataframes of tide data.}

\item{datetimecol}{The column name containing timestamps.}

\item{heightcol}{The column name containing water levels.}

\item{sigmacol}{The column name containing sigma values.}
}
\value{
A single dataframe of tide data.
}
\description{
Combine data files produced by \code{download_tides}.
}
\seealso{
download_tides
}
