% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rreformat.r
\name{merge_waves}
\alias{merge_waves}
\title{Combine Wave Records}
\usage{
merge_waves(waves)
}
\arguments{
\item{waves}{A list of dataframes of wave data.}
}
\value{
A single dataframe of wave data.
}
\description{
Combine data files produced by \code{download_waves}.
}
\details{
The wave data also contains meteorological data
  such as pressure, air temperature, sea surface temperature,
  etc. The wave data records can be very spotty.
}
\seealso{
download_waves
}
