% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rreadwrite.r
\name{read_bathymetry_tables}
\alias{read_bathymetry_tables}
\title{Read Bathymetry Tables}
\usage{
read_bathymetry_tables(files)
}
\arguments{
\item{files}{A list of files containing bathymetry data. Each file is 
expected to have columns \code{OID}, \code{dist}, \code{count}, 
\code{area}, \code{sum}, and \code{bedelev}.}
}
\value{
A data frame with columns, \code{bedelev}, \code{dist}, \code{area}, 
  \code{count}, and \code{sum}.
}
\description{
Read and combine the bathymetry tables.
}
