% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rreadwrite.r
\name{read_rrectd}
\alias{read_rrectd}
\title{Read CTD Data}
\usage{
read_rrectd(fpath, quiet = FALSE, distcol = "dist",
  surfdistcol = "surfdist")
}
\arguments{
\item{fpath}{Path to input data.}

\item{quiet}{Suppress messages during data processing.}

\item{distcol}{Column name containing distance upstream for each cast
location.}

\item{surfdistcol}{Column name containing water surface elevation for each
cast location.}
}
\value{
A dataframe with named columns. Additional elements (e.g. note) are
  added as attributes.
}
\description{
Read CTD transect data from Largier Lab Matlab file.
}
\details{
Largier Lab Matlab data is typically formatted as a Matlab
  structure with multiple equal-length data columns and two columns
  'surfdist' and 'surfelev' with surface information for each cast location.
  'Surfdist' and 'surfelev' are used to populate an expanded 'surfelev'
  column corresponding to data column 'elev'.
}

