% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rreadwrite.r
\name{read_rrectdgrid}
\alias{read_rrectdgrid}
\title{Read interpolated CTD grid}
\usage{
read_rrectdgrid(fpath, quiet = FALSE)
}
\arguments{
\item{fpath}{Path to input data.}

\item{quiet}{Suppress messages during data processing.}
}
\value{
A dataframe with named columns. Additional elements (e.g. note) are
  added as attributes.
}
\description{
Read interpolated grid of CTD data from Largier Lab Matlab file.
}
\details{
Largier Lab Matlab data is typically formatted as a Matlab  
  structure with multiple equal-size matrices.
}
