% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rreadwrite.r
\name{read_rrewll}
\alias{read_rrewll}
\title{Read Water Level Data}
\usage{
read_rrewll(fpath, quiet = FALSE)
}
\arguments{
\item{fpath}{Path to input data.}

\item{quiet}{Suppress messages during data processing.}
}
\value{
A dataframe with named columns. Additional elements (e.g. note) are
  added as attributes.
}
\description{
Read water level data from Largier Lab Matlab file
}
\details{
Largier Lab Matlab data is typically formatted as a Matlab  
  structure with multiple equal-length data columns and an optional note.
}
