% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrexplore.r
\name{stratify}
\alias{stratify}
\title{Stratify by a Variable}
\usage{
stratify(x, midpoints, ...)
}
\arguments{
\item{x}{Vector of values.}

\item{midpoints}{Vector of midpoints or cut points to stratify \code{x} by.
\code{min(x)} and \code{max(x)} are automatically included.}

\item{...}{Other arguments to pass to \code{cut()}.}
}
\description{
Stratify data into zones using specified midpoints.
}
\details{
This function is basically a wrapper for \code{cut()}, but requires
  the cuts be directly specified and that \code{include.lowest = TRUE}.
  Every element of \code{x} is therefore guaranteed to be included in one
  interval.
}

