% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rrexplore.r
\name{stratify_depth}
\alias{stratify_depth}
\title{Stratify by Depth}
\usage{
stratify_depth(z, midpoints, ...)
}
\arguments{
\item{z}{Vector of depths}

\item{midpoints}{Vector of midpoints or cut points to stratify \code{z} by.
\code{min(z)} and \code{max(z)} are automatically included.}

\item{...}{Other arguments to pass to \code{cut()}.}
}
\description{
Stratify depths into zones using specified midpoints.
}
\details{
This function is basically a wrapper for \code{cut()}, but requires
  the cuts be directly specified and that \code{include.lowest = TRUE}. It
  Every element of \code{z} is therefore guaranteed to be included in one
  interval.
}

