% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrevaluate.r
\name{summarize_by_strata}
\alias{summarize_by_strata}
\title{Summarize by Strata}
\usage{
summarize_by_strata(d, stratcols, summexpr)
}
\arguments{
\item{d}{The dataframe to be summarized.}

\item{stratcols}{Vector or list of column names to group by.}

\item{summexpr}{Named vector or list of expressions to summarize the data.}
}
\value{
A dataframe containing grouped and summarized data.
}
\description{
Group and summarize a dataframe by specific "strata" or groupings.
}
\details{
This function is basically a wrapper for the \code{summarize} and
  \code{group_by} functions provided by \code{dplyr}.
}
