% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rrevaluate.r
\name{add_volume}
\alias{add_volume}
\title{Calculate Volume}
\usage{
add_volume(d, v, r, zcol = "elev", xcol = "dist")
}
\arguments{
\item{d}{Dataframe containing physical data, e.g. \code{data(grids)}.}

\item{v}{Dataframe containing volume lookup information, e.g.
\code{data(volumes)}.}

\item{zcol}{Column name containing vertical markers, e.g. \code{elev}.}

\item{xcol}{Column name containing longitudinal markers, e.g. \code{dist}.}
}
\value{
The dataframe \code{d} with the additional column \code{volume}.
}
\description{
Calculate water volume contained in depth and distance zones.
}

